YAFS-AOS - fix jsonify function, add power ppi, test for device creation

+++ src/fogExtension/devices/GenericHardware.py
@@ -8,7 +8,7 @@ class GenericDevice:
         Generic definition of mandatory parameters and power model.
 
         Args:
-                freq (int): frequency of the device in MHz,
+                freq (float): frequency of the device in MHz,
                 ram (int): memory size of the device in MB, it is a constant set at 4 (default value),
                 pm (PowerModel) : defines the power-model of the device [Wired, Battery, Solar],
         """
